+++ stellar-contract-env-host/src/host/err_helper.rs
